# Conflicts:
#	src/commands/prepare/index.ts
#	src/common/command.abstract.ts

+++ src/commands/compile.integration.test.ts
@@ -43,7 +43,9 @@ describe('Compile CLI command', () => {
   });
 
   afterEach(async () => {
-    await rimraf(tempDir);
+    if (process.cwd() == '') {
+      await rimraf(tempDir);
+    }
   });
 
   describe('integration tests', () => {
